// src/types/navigation.types.ts
export interface DocsDropdownItem {
  titleEs: string;
  titleEn: string;
  descEs: string;
  descEn: string;
  href: string;
  hrefEn: string;
}

export interface RecipeDropdownItem {
  id: string;
  nameEs: string;
  nameEn: string;
  descEs: string;
  descEn: string;
  icon: string;
  shadow: string;
  color: string;
}

export interface FaqCarouselSlide {
  questionEs: string;
  questionEn: string;
  answerEs: string;
  answerEn: string;
}